function append(delete)

+++ accounts/views.py
@@ -53,9 +53,10 @@ def signup(request):
 
 
 def delete(request):
-    if request.user.is_authenticated:
-        request.user.delete()
-        auth_logout(request)
+    if request.method == 'POST':
+        if request.user.is_authenticated:
+            request.user.delete()
+            auth_logout(request)
     return redirect('movies:index')
